# 8298865: Excessive memory allocation in CipherOutputStream AEAD decryption