Improves `CipherOutputStream` with `AES/GCM` significantly.

Additional testing:
 - [ ] MacOS AArch64 server fastdebug, `jdk_security`
 - [ ] Linux x86_64 server fastdebug, `all`
 - [x] Ad-hoc benchmarks 

```
Benchmark                        (dataSize)  (keyLength)    Mode  Cnt       Score     Error  Units

# Before
Benchmark                         (dataSize)  (keyLength)   Mode  Cnt       Score      Error  Units
AESGCMCipherOutputStream.decrypt       16384          128  thrpt   40   37119.222 ±  115.021  ops/s
AESGCMCipherOutputStream.decrypt     1048576          128  thrpt   40      17.100 ±    0.036  ops/s
AESGCMCipherOutputStream.encrypt       16384          128  thrpt   40  173869.510 ± 1077.892  ops/s
AESGCMCipherOutputStream.encrypt     1048576          128  thrpt   40    2850.846 ±   19.319  ops/s

# After
AESGCMCipherOutputStream.decrypt       16384          128  thrpt   40   70632.205 ±  124.632  ops/s
AESGCMCipherOutputStream.decrypt     1048576          128  thrpt   40     794.142 ±  301.196  ops/s
AESGCMCipherOutputStream.encrypt       16384          128  thrpt   40  171715.301 ± 1157.267  ops/s
AESGCMCipherOutputStream.encrypt     1048576          128  thrpt   40    2827.844 ±   28.430  ops/s
```

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue
- [ ] [JDK-8298865](https://bugs.openjdk.org/browse/JDK-8298865) needs maintainer approval

### Integration blocker
&nbsp;⚠️ Dependency #2393 must be integrated first

### Issue
 * [JDK-8298865](https://bugs.openjdk.org/browse/JDK-8298865): Excessive memory allocation in CipherOutputStream AEAD decryption (**Bug** - P4)


### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.org/jdk17u-dev.git pull/2394/head:pull/2394` \
`$ git checkout pull/2394`

Update a local copy of the PR: \
`$ git checkout pull/2394` \
`$ git pull https://git.openjdk.org/jdk17u-dev.git pull/2394/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 2394`

View PR using the GUI difftool: \
`$ git pr show -t 2394`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.org/jdk17u-dev/pull/2394.diff">https://git.openjdk.org/jdk17u-dev/pull/2394.diff</a>

</details>