@shipilev 2024-04-11
The dependent PR cannot be integrated, closing this one as well.